requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
django-cors-headers==4.3.1
Pillow==10.1.0
numpy==1.25.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0